requirements.txt
flask
celery
pyyaml
cerberus
requests